Center grid

+++ WFC-2D/Assets/GridManager.cs
@@ -41,7 +41,7 @@ public class GridManager : MonoBehaviour
         float gridW = cols * tileSize;
         float gridH = rows * tileSize;
 
-        transform.position = new Vector2(-gridW / 2 + tileSize / 2, gridH / 2 - tileSize / 2);
+        transform.position = new Vector2(-gridW / 2 + (float) tileSize / 2, gridH / 2 - (float) tileSize / 2);
     }
 
     // Update is called once per frame
